refactor: split errors module
